+++ src/model/Transactions.js
@@ -3,7 +3,7 @@ const { Model, DataTypes } = require('sequelize');
 class Transactions extends Model {
   static init(connection) {
     super.init({
-      amount: DataTypes.DECIMAL(10, 2),
+      amount: DataTypes.INTEGER,
       account_payer_id: DataTypes.STRING,
       account_receiver_id: DataTypes.STRING,
     }, {
